Add User & Remember Me Migration
Add both the User and Remember Me migration together since remember me
will be needed later on to remember the users login.

Also for the User model instead of an email address to login it will use
the user name since this will be used to display which user posted what
and to give some anonymity to the post.