Change Sign up & Login \ out Test Start From Root
In order to have a more realized test change the start URL from going
directly from the page to where the user would start thus if there is
any errors in any of the process it will display.

Also change Root from string to use proper route prefixes in case it
changes it in the future.