Change Post In Route To Be Simplify
Because their will always be a user that adds, edits or destroys a post
and there is any error when after fixing validation errors. Which will
result in having more then one path to a post and make the routes more
complex. Remove all unnecessary paths to the post and simplify them.