Move Daily Reminder Timer JS Selector Inside If
Move the Daily Reminder Timer JS selector from
checkEnableDailyReminder() to be inside the if statement in order not to
run until there is an enable daily reminder is found on the page. This
will speed up loading of the page since there is no pointless selection
until it is necessary.