Move Render Markdown From Post View To Controller
Move the method call render_markdown_post_to_html from Post view to its
Controller because this is proper MVC design pattern since the
controller is the one that calls methods from the model and displays the
results to the view. In this case render_markdown_post_to_html is being
called from the view instead of the controller which is not proper.