Remove New Action Method From Comments Controller
Remove the New Action Method from the Comments controller since the
comment form needs to be in Post show page. Because of this, @comment
will be supply to by the Posts Controller instead of the form is being
rendered in its views.

Also, since the New Action Method is no longer in the Comments
controller its routes has been remove to keep the routes clean.