Remove Edit Action Method From Comments Controller
Remove the Edit Action Method from the Comments controller since the
comment form needs to be in Post show page. Because of this, @comment
will be supply to by the Posts Controller since the form is being
rendered in its views.

Also, since the Edit Action Method is no longer in the Comments
controller its routes has been remove to keep the routes clean.