Remove Index Action Method From Comments CRTL
Remove the Index Action Method from the Comments controller since the
comments needs to be in Post show page. Because of this, @comments will
be supply to by the Posts Controller and Comments Create Action Method
if the page errors out since the index is being rendered in Posts show
view.

Also, since the Index Action Method is no longer in the Comments
controller its routes has been remove to keep the routes clean.