Add Publish Date To Post Test Factory
Add Publish Date to the Post test factory in order to test the publish
date. Also it is in an array in order to have a mix of nil or date as
the actully site will have both on it the test should be able to filter
out nil pubilsh dates and only display the current postings.