Change Profile Route To Allow Any User's Profile
Change the profile route to allow any user's profile because the profile
view is meant to showcase the activated of the user on the application.
If it is only for the current user being logged in then this defeats
the propose of having a show view to begin with.

Update the controller and link to the profile in the main header to use
the new route and id that came from it.